#AB19 comment fix

+++ src/main/java/com/healthassist/service/AdminService.java
@@ -155,24 +155,24 @@ public class AdminService {
                 doctorAppointmentRepository.deleteByPatientId(user.getUserId());
                 patientRecordRepository.deleteByPatientId(user.getUserId());
             }
-		/* if (authorityName == AuthorityName.ROLE_COUNSELOR) {
-			List<CounselorAppointment> appointments = counselorAppointmentRepository.findByCounselorId(user.getUserId());
-			for (CounselorAppointment appointment : appointments) {
-				List<PatientRecord> patientRecords = patientRecordRepository.findByAppointmentIdAndStatus(appointment.getAppointmentId(), PatientRecordStatus.COUNSELOR_APPOINTMENT);
-				patientRecords.forEach(patientRecord -> {
-					if (patientRecord.getActivePatientId() != null)
-						activePatientRepository.deleteByActivePatientId(patientRecord.getActivePatientId());
-					patientRecordRepository.deleteByPatientRecordId(patientRecord.getPatientRecordId());
-				});
-				patientRecords = patientRecordRepository.findByAppointmentIdAndStatus(appointment.getAppointmentId(), PatientRecordStatus.COUNSELOR_IN_PROGRESS);
-				patientRecords.forEach(patientRecord -> {
-					if (patientRecord.getActivePatientId() != null)
-						activePatientRepository.deleteByActivePatientId(patientRecord.getActivePatientId());
-					patientRecordRepository.deleteByPatientRecordId(patientRecord.getPatientRecordId());
-				});
-				counselorAppointmentRepository.deleteByAppointmentId(appointment.getAppointmentId());
-			}
-		}*/
+             if (authorityName == AuthorityName.ROLE_COUNSELOR) {
+                List<CounselorAppointment> appointments = counselorAppointmentRepository.findByCounselorId(user.getUserId());
+                for (CounselorAppointment appointment : appointments) {
+                    List<PatientRecord> patientRecords = patientRecordRepository.findByAppointmentIdAndStatus(appointment.getAppointmentId(), PatientRecordStatus.COUNSELOR_APPOINTMENT);
+                    patientRecords.forEach(patientRecord -> {
+                        if (patientRecord.getActivePatientId() != null)
+                            activePatientRepository.deleteByActivePatientId(patientRecord.getActivePatientId());
+                        patientRecordRepository.deleteByPatientRecordId(patientRecord.getPatientRecordId());
+                    });
+                    patientRecords = patientRecordRepository.findByAppointmentIdAndStatus(appointment.getAppointmentId(), PatientRecordStatus.COUNSELOR_IN_PROGRESS);
+                    patientRecords.forEach(patientRecord -> {
+                        if (patientRecord.getActivePatientId() != null)
+                            activePatientRepository.deleteByActivePatientId(patientRecord.getActivePatientId());
+                        patientRecordRepository.deleteByPatientRecordId(patientRecord.getPatientRecordId());
+                    });
+                    counselorAppointmentRepository.deleteByAppointmentId(appointment.getAppointmentId());
+                }
+            }
             if (authorityName == AuthorityName.ROLE_DOCTOR) {
                 List<DoctorAppointment> appointments = doctorAppointmentRepository.findByDoctorId(user.getUserId());
                 for (DoctorAppointment appointment : appointments) {
